+++ tests/test_autodoc_skip.py
@@ -3,7 +3,7 @@ import pytest
 
 @pytest.mark.sphinx("html", testroot="docstrings")
 def test_exclude_members(app, do_autodoc):
-    options = {"members": True, "undoc-members": True}
+    options = {"members": None, "undoc-members": None}
     assert not any(
         "base_fields" in line
         for line in do_autodoc(app, "module", "dummy_django_app.forms", options)
@@ -12,7 +12,7 @@ def test_exclude_members(app, do_autodoc):
 
 @pytest.mark.sphinx("html", testroot="docstrings")
 def test_include_members(app, do_autodoc):
-    options = {"members": True}
+    options = {"members": None}
     assert any(
         "__init__" in line
         for line in do_autodoc(app, "module", "dummy_django_app.forms", options)