+++ CHANGES.rst
@@ -7,6 +7,7 @@ UNRELEASED
 * Add support for Python 3.9
 * Add support for django-mptt with Django >=3.1
 * Append initial docstrings to attributes
+* Fix mutable references of pre-commit hooks
 
 
 Version 1.0.2 (2021-02-02)