+++ CHANGES.rst
@@ -8,6 +8,7 @@ UNRELEASED
 * Add support for django-mptt with Django >=3.1
 * Append initial docstrings to attributes
 * Fix mutable references of pre-commit hooks
+* Fix tests for sphinx 3.5.0
 
 
 Version 1.0.2 (2021-02-02)